+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ui/storybook
 
+## 0.0.84
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/react@0.0.82
+
+## 0.0.83
+
+### Patch Changes
+
+- @jamsr-ui/react@0.0.81
+
 ## 0.0.82
 
 ### Patch Changes