+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/select
 
+## 0.0.37
+
+### Patch Changes
+
+- fix
+
+## 0.0.36
+
+### Patch Changes
+
+- fix
+
 ## 0.0.35
 
 ### Patch Changes